change the test_message_views

+++ app.py
@@ -312,6 +312,12 @@ def messages_destroy(message_id):
         return redirect("/")
 
     msg = Message.query.get(message_id)
+
+    if g.user.id != msg.user_id:
+        #can only delete your own message
+        flash("Access unauthorized.", "danger")
+        return redirect("/")
+
     db.session.delete(msg)
     db.session.commit()
 
@@ -327,9 +333,12 @@ def messages_add_like(message_id):
 
     msg = Message.query.get_or_404(message_id)
 
+    if g.user.id == msg.user_id:
+        flash("Can not like your own message!" , "danger")
+        return redirect('/')
+
     if g.user.is_like(msg):
         # alrealy like this message, unlike it
-
         Likes.query.filter(Likes.user_id==g.user.id , Likes.message_id==msg.id).delete()
         db.session.commit()
